Updated SG10K example (DAG and cmd)

+++ SG10K/example_dag_and_cmds.sh
@@ -4,18 +4,28 @@ set -euo pipefail;
 wrapper=$(readlink -f $(dirname $0)/SG10K.py)
 outdir=$(dirname $wrapper)
 
-FASTQ_DIR="$RPD_ROOT/testing/data/SG10K/illumina-platinum-NA12878/"
-test -d $FASTQ_DIR
-fastqs_1=$(ls $FASTQ_DIR/*ERR091571_1*1M-only.fastq.gz)
-fastqs_2=$(ls $FASTQ_DIR/*ERR091571_2*1M-only.fastq.gz)
+CFG=$RPD_ROOT/testing/data/SG10K/MUX3275-WHH474.yaml
+test -e $CFG
 
 tmpdir=$(mktemp -d)
-rm -rf $tmpdir
-$wrapper -1 $fastqs_1 -2 $fastqs_2 --sample example -o $tmpdir --no-run
+rmdir $tmpdir
+echo "DEBUG: tmpdir is $tmpdir" 1>&2
+$wrapper -c $CFG --sample WHH474 -o $tmpdir --no-run
 
+pushd $tmpdir;
+dag=$outdir/example_workflow_graph.pdf
+echo "Writing DAG to $dag"
+EXTRA_SNAKEMAKE_ARGS="--dryrun --quiet --dag" bash run.sh;
+cat logs/snakemake.log | dot -Tpdf > $dag
+cmds=$outdir/example_workflow_commands.txt
+echo "Writing cmds to $cmds"
+EXTRA_SNAKEMAKE_ARGS="--dryrun --quiet --printshellcmds" bash run.sh;
+cut -f 2 -d ']' logs/snakemake.log > $cmds
+popd
 
-cd $tmpdir;
-EXTRA_SNAKEMAKE_ARGS="--dryrun --quiet --dag" bash run.sh  | dot -Tpdf > $outdir/example_workflow_graph.pdf 
-EXTRA_SNAKEMAKE_ARGS="--dryrun --quiet --printshellcmds" bash run.sh | cut -f 2 -d ']' > $outdir/example_workflow_commands.txt
+if true; then
+	echo "DEBUG: not deleting $tmpdir" 1>&2;
+else
+	test -d $tmpdir && rm -rf $tmpdir;
+fi
 
-test -d $tmpdir && rm -rf $tmpdir
